Vulkan Driver Errata Boostrap
This repository documents known driver or platform issues, how they are
triggered and how they can be worked around.  This database of bugs is
machine readable and found in the `errata/` directory.  Baked in this
directory is also generated C header and source files (under the `src/`
directory) that provide an easy way for applications to know which bugs
need to be worked around on the running device.